requirements.txt
python-editor
PyYAML
requests
